minor ui tweaks
