+++ src/components/assignment/AssignmentQuestions.tsx
@@ -6,60 +6,7 @@ import assert from "assert";
 import { canUserAccessAssignment } from "@/utils/classroom/canUserAccessAssignment";
 import { EnumAccessType } from "@/schemas/dbTableAccessSchema";
 import ChooseQuestionTypeDialog from "../question/add/ChooseQuestionTypeDialog";
-import { type QuestionType } from "@/schemas/questionSchema";
-import { EditQuestionDataWrapper } from "../question/edit/EditQuestionDataWrapper";
-import {
-  Accordion,
-  AccordionContent,
-  AccordionItem,
-  AccordionTrigger,
-} from "@/components/ui/accordion";
-import { questionDisplayConfigByType } from "@/utils/constants/misc";
-import { Badge } from "@/components/ui/badge";
-import { cn } from "@/utils/cn";
-
-type QuestionTitleProps = {
-  classroomId: string;
-  assignmentId: string;
-  questionId: string;
-  name: string;
-  type: QuestionType;
-};
-
-function QuestionTitle({ type, questionId, name }: QuestionTitleProps) {
-  const displayConfig = questionDisplayConfigByType[type];
-
-  return (
-    <Accordion type="multiple" className="w-full border-x border-border">
-      <AccordionItem value={questionId}>
-        <div className="flex w-full items-center justify-between">
-          <div className="w-full">
-            <AccordionTrigger className="grow">
-              <div className="flex flex-col gap-2 pl-4">
-                <p>{name}</p>
-                <Badge
-                  className={cn(
-                    "flex items-center gap-2",
-                    displayConfig.badgeStyles
-                  )}
-                >
-                  {displayConfig.icon}
-                  {displayConfig.label}
-                </Badge>
-              </div>
-            </AccordionTrigger>
-          </div>
-          <EditQuestionDataWrapper type={type} id={questionId}>
-            <Button variant="ghost">
-              <PencilIcon className="h-4 w-4" />
-            </Button>
-          </EditQuestionDataWrapper>
-        </div>
-        <AccordionContent>Question Here</AccordionContent>
-      </AccordionItem>
-    </Accordion>
-  );
-}
+import { QuestionTitle } from "./QuestionTitle";
 
 type AssignmentQuestionsProps = {
   classroomId: string;
@@ -100,11 +47,12 @@ export async function AssignmentQuestions({
   }
 
   return (
-    <ol className="flex flex-col">
+    <ol className="flex flex-col gap-3">
       {questions.map(({ id, name, type }) => {
         return (
           <li className="flex flex-row items-center" key={id}>
             <QuestionTitle
+              isAuthorizedToAddOrDelete={isAuthorizedToAddOrDelete}
               classroomId={classroomId}
               assignmentId={assignmentId}
               questionId={id}
